requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
colorama==0.4.1
decorator==4.4.0
Django==2.2.1
django-cors-headers==3.0.1
djangorestframework==3.9.4
ipython==7.5.0
ipython-genutils==0.2.0
jedi==0.13.3
more-itertools==7.0.0
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
pluggy==0.11.0
prompt-toolkit==2.0.9
psycopg2==2.8.2
ptyprocess==0.6.0
py==1.8.0
Pygments==2.4.0
pytest==4.5.0
pytest-django==3.4.8
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.2
wcwidth==0.1.7
